refactor: move prisma extension to own file
Importing other functions from the same file will leak server code into remix frontend. I forgot this!
